+++ app/cont/views.py
@@ -1,6 +1,6 @@
 from datetime import date, datetime
 from flask import flash, render_template, session, redirect, request, url_for
-from flask_login import login_required
+from flask_login import current_user, login_required
 from . import cont
 from .forms import ContactForm
 from .. import db
@@ -10,6 +10,8 @@ from ..models import User, Contact
 @cont.route("/contacts/<int:contact_id>/delete", methods=["GET", "POST"])
 @login_required
 def delete_contact(contact_id):
+    if current_user != task.user_id and not current_user.can(Permission.ADMIN):
+        abort(403)
     contact = Contact.query.get_or_404(contact_id)
     db.session.delete(contact)
     db.session.commit()
@@ -20,8 +22,10 @@ def delete_contact(contact_id):
 @cont.route("/contacts", methods=["GET"])
 @login_required
 def contacts():
-    contacts = Contact.query.order_by(Contact.first).all()
-    return render_template("cont/contacts.html", title="Contacts", contacts=contacts)
+    if current_user.is_authenticated:
+        contacts = Contact.query.filter_by(user_id=current_user.id).all()
+        #.order_by(Contact.first).all()
+        return render_template("cont/contacts.html", title="Contacts", contacts=contacts)
 
 
 @cont.route("/contacts/<int:contact_id>", methods=["GET"])
@@ -45,6 +49,7 @@ def new_contact():
             last_contact=form.last_contact.data,
             next_contact=form.next_contact.data,
             notes=form.notes.data,
+            user_id=current_user._get_current_object(),
         )
         db.session.add(contact)
         db.session.commit()
@@ -57,6 +62,8 @@ def new_contact():
 @cont.route("/contacts/<int:contact_id>/update", methods=["GET", "POST"])
 @login_required
 def update_contact(contact_id):
+    if current_user != task.user_id and not current_user.can(Permission.ADMIN):
+        abort(403)
     contact = Contact.query.get_or_404(contact_id)
     form = ContactForm(contact=contact)
     if form.validate_on_submit():
